+++ lib/ffi_generated.mli
@@ -1,2 +1,3 @@
 type 'a fn = 'a
 val foreign : string -> ('a -> 'b) Ctypes.fn -> ('a -> 'b)
+val foreign_value : string -> 'a Ctypes.typ -> 'a Ctypes.ptr